requirements.txt
discord.py
python-dotenv
anthropic
openai
redis
